updating map render
